Correção de erros de texto e adição de comentários

+++ app/crud_Prova.py
@@ -34,25 +34,31 @@ def menu_prova():
         else:
             print("Opção inválida.")
 
+        input("Aperte Enter para continuar")
+
 def adicionar_prova():
     try:
+        # Pega os dados da nova prova
         nota_max = int(input("Nota máxima: "))
         qtd_questoes = int(input("Quantidade de questões: "))
         data_str = input("Data de criação (AAAA-MM-DD): ")
         tipo = input("Tipo (Objetiva, Dissertativa, Mista): ")
-        professor_id = input("ID do professor: ")
+        professor_id = input("ID do(a) professor(a): ")
         disciplina_id = input("ID da disciplina: ")
 
         data_criacao = datetime.strptime(data_str, "%Y-%m-%d").date()
 
+        # Conecta ao BD e cria cursor
         conn = conectar()
         cur = conn.cursor()
 
+        # Faz a inserção do novo registro
         cur.execute("""
             INSERT INTO Prova (nota_max, qtd_questoes, data_criacao, tipo, fk_Professor_id, fk_Disciplina_id)
             VALUES (%s, %s, %s, %s, %s, %s)
         """, (nota_max, qtd_questoes, data_criacao, tipo, professor_id, disciplina_id))
 
+        # Commita, encerrando a operação
         conn.commit()
         print("Prova cadastrada com sucesso!")
 
@@ -61,26 +67,31 @@ def adicionar_prova():
 
     finally:
         if conn:
+            # Encerra conexão com o BD e cursor
             cur.close()
             conn.close()
 
 
 def editar_prova():
     try:
+        # Pega o id da prova a ser editada
         id_prova = input("ID da prova a editar: ")
 
+        # Pega os novos dados da prova
         nota_max = int(input("Nova nota máxima: "))
         qtd_questoes = int(input("Nova quantidade de questões: "))
         data_str = input("Nova data de criação (AAAA-MM-DD): ")
         tipo = input("Novo tipo: ")
-        professor_id = input("Novo ID do professor: ")
+        professor_id = input("Novo ID do(a) professor(a): ")
         disciplina_id = input("Novo ID da disciplina: ")
 
         data_criacao = datetime.strptime(data_str, "%Y-%m-%d").date()
 
+        # Cria conexão com o BD e cursor
         conn = conectar()
         cur = conn.cursor()
 
+        # Realiza as mudanças no registro da prova
         cur.execute("""
             UPDATE Prova
             SET nota_max = %s, qtd_questoes = %s, data_criacao = %s, tipo = %s,
@@ -88,6 +99,7 @@ def editar_prova():
             WHERE id = %s
         """, (nota_max, qtd_questoes, data_criacao, tipo, professor_id, disciplina_id, id_prova))
 
+        # Commita as mudanças e encerra a operação
         conn.commit()
         print("Prova atualizada com sucesso!")
 
@@ -96,17 +108,27 @@ def editar_prova():
 
     finally:
         if conn:
+            # Encerra conexão com o BD e cursor
             cur.close()
             conn.close()
 
 
 def excluir_prova():
     try:
+        # Pega o id da prova a ser excluída
         id_prova = input("ID da prova a excluir: ")
 
+        # Cria conexão com o BD e cria o cursor
         conn = conectar()
         cur = conn.cursor()
 
+        # Verifica se a prova existe no banco
+        cur.execute("SELECT * FROM Prova WHERE id = %s", (id_prova,))
+        if cur.fetchone() is None:
+            print("Prova não encontrada.")
+            return
+
+        # Deleta a prova
         cur.execute("DELETE FROM Prova WHERE id = %s", (id_prova,))
         conn.commit()
         print("Prova excluída com sucesso!")
@@ -116,15 +138,18 @@ def excluir_prova():
 
     finally:
         if conn:
+            # Finaliza a conexão com o BD e o cursor
             cur.close()
             conn.close()
 
 
 def listar_todas_as_provas():
     try:
+        # Conecta ao Bd e cria cursor
         conn = conectar()
         cur = conn.cursor()
 
+        # Pega dos os registros de prova
         cur.execute("""
             SELECT p.id, p.tipo, p.data_criacao, p.nota_max, d.nome AS disciplina, pr.nome AS professor
             FROM Prova p
@@ -133,6 +158,7 @@ def listar_todas_as_provas():
             ORDER BY p.id
         """)
 
+        # Salva todos os registros
         provas = cur.fetchall()
 
         print("\nProvas cadastradas:")
@@ -144,17 +170,35 @@ def listar_todas_as_provas():
 
     finally:
         if conn:
+            # Finaliza conexão com o BD e cursor
             cur.close()
             conn.close()
 
 
 def listar_provas_por_professor():
     try:
+        # Pega o id do professor
         id_professor = input("ID do professor: ")
 
+        # Conecta ao BD e cria cursor
         conn = conectar()
         cur = conn.cursor()
 
+        # Pega o nome do professor
+        cur.execute("""
+            SELECT nome
+            FROM Professor
+            WHERE id = %s
+        """, (id_professor,))
+        
+        professor_nome = cur.fetchone()
+        
+        # Verifica se o professor existe
+        if professor_nome is None:
+            print("Professor(a) não encontrado.")
+            return
+
+        # Seleciona as provas que foram criadas pelo determinado professor
         cur.execute("""
             SELECT p.id, p.tipo, p.data_criacao, p.nota_max, d.nome
             FROM Prova p
@@ -165,7 +209,7 @@ def listar_provas_por_professor():
 
         provas = cur.fetchall()
 
-        print(f"\nProvas do professor {id_professor}:")
+        print(f"\nProvas do(a) professor(a) {professor_nome[0]}:")
         for p in provas:
             print(f"ID: {p[0]}, Tipo: {p[1]}, Data: {p[2]}, Nota Máx: {p[3]}, Disciplina: {p[4]}")
 
@@ -174,17 +218,22 @@ def listar_provas_por_professor():
 
     finally:
         if conn:
+            # Encerra conexão com BD e cursor
             cur.close()
             conn.close()
 
 
+
 def listar_prova_por_id():
     try:
+        # Pega o ID da prova
         id_prova = input("ID da prova: ")
 
+        # Conecta ao BD e cria cursor
         conn = conectar()
         cur = conn.cursor()
 
+        # Busca a prova desejada
         cur.execute("""
             SELECT p.id, p.tipo, p.data_criacao, p.nota_max, d.nome, pr.nome
             FROM Prova p
@@ -205,17 +254,35 @@ def listar_prova_por_id():
 
     finally:
         if conn:
+            # Encerra a conexão e o cursor
             cur.close()
             conn.close()
 
 
 def listar_provas_por_disciplina():
     try:
+        #Pega o Id da dsiciplina
         id_disc = input("ID da disciplina: ")
 
+        # Faz a conexão com o BD e cria o cursor
         conn = conectar()
         cur = conn.cursor()
 
+        # Busca o nome da disciplina
+        cur.execute("""
+            SELECT nome
+            FROM Disciplina
+            WHERE id = %s
+        """, (id_disc,))
+        
+        disciplina_nome = cur.fetchone()
+
+        # Verifica se a disciplina está registrada no BD
+        if disciplina_nome is None:
+            print("Disciplina não encontrada.")
+            return
+
+        # Seleciona as provas da determinada disciplina
         cur.execute("""
             SELECT p.id, p.tipo, p.data_criacao, p.nota_max, pr.nome
             FROM Prova p
@@ -226,7 +293,7 @@ def listar_provas_por_disciplina():
 
         provas = cur.fetchall()
 
-        print(f"\nProvas da disciplina {id_disc}:")
+        print(f"\nProvas da disciplina {disciplina_nome[0]}:")
         for p in provas:
             print(f"ID: {p[0]}, Tipo: {p[1]}, Data: {p[2]}, Nota Máx: {p[3]}, Professor: {p[4]}")
 
@@ -235,5 +302,6 @@ def listar_provas_por_disciplina():
 
     finally:
         if conn:
+            # Finaliza conexão e cursor
             cur.close()
             conn.close()
